refactor: Move GuaranteedTransfer to OTP model
